TTHAnalysis: further refactoring of analyzers for susy
